requirements.txt

Babel==0.9.6
#Cython==0.19.1
Django==1.5.4
Flask==0.10.1
Flask-S3==0.1.3
Flask-WTF==0.8.2
Jinja2==2.7
MarkupSafe==0.18
Pillow==1.7.8
#Py-==1.0.0
PyYAML==3.10
Pygments==1.6
South==0.7.6
#Sphinx==1.2b2
#Tkinter==0.0.0
Unidecode==0.04.14
WTForms==1.0.2
Wand==0.3.5
WebOb==1.2.3
WebTest==2.0.9
Werkzeug==0.9.1
Whoosh==2.5.4
beautifulsoup4==4.3.1
blessings==1.5.1
boto==2.8.0
catkin-pkg==0.1.18
coverage==3.6
coveralls==0.1.1
detox==0.9.2
dj-static==0.0.5
django-appconf==0.6
django-cache-panel==0.1
django-compressor==1.3
django-crispy-forms==1.4.0
#-e git://github.com/dcramer/django-debug-toolbar.git@bf39bb03a0777505f1fbca2d05f31db4aef37a46#egg=django_debug_toolbar-dev
django-debug-toolbar-template-timings==0.5.1
django-dynamic-fixture==1.6.5
django-extensions==0.9
django-extra-views==0.5.5
django-haystack==2.1.0
django-heroku-memcacheify
django-jinja==0.21
#django-nose==1.1
#-e git+git@github.com:rsbusby/django-oscar.git@c1d3be31973966de4153a0f60a450f32ffdbc0da#egg=django_oscar-dev
django-rosetta==0.6.8
django-s3-folder-storage
django-treebeard==2.0b2
djrill==0.6.0
#django-webtest==1.5.7
#docopt==0.4.0
#docutils==0.11
#empy
easypost==2.0.6
eventlet==0.14.0
flake8==2.0
#flann
flask-mongoengine==0.6
geopy==0.95.1
greenlet==0.4.1
gunicorn==18.0
#ipdb==0.7.1
ipython==0.13.2
itsdangerous==0.22
mandrill==1.0.27
mccabe==0.2.1
#mercurial==2.7
#mock==1.0.1
mongoengine==0.8.2
#nose==1.2.1
#nose-progressive==1.5
#pep8==1.4.6
psycopg2==2.5.1
purl==0.8
#py==1.4.15
#pycairo==1.10.0
pyelasticsearch
pyflakes==0.7.3
#pygccxml==1.1.0
pylibmc==1.2.3
pymongo==2.5.2
pyprof2calltree==1.1.0
python-dateutil==1.5
python-magic==0.4.3
python-memcached==1.51
python-social-auth==0.1.21
#pytz==2013b
rauth==0.5.4
#readline==6.2.4.1 ## this takes forever on Heroku
requests==1.1.0
roman==2.0.0
rope==0.9.2
#rosdep==0.10.21
#rosdistro==0.2.11
#rosinstall==0.6.29
#rospkg==1.0.21
#sh==1.09
#six==1.4.1
sorl-thumbnail==11.12
#spec==0.11.1
#static==0.4
stripe==1.7.7
#tox==1.6.1
#transifex-client==0.8
#vcstools==0.1.31
#virtual-less==0.0.1-1.3.3
#virtual-node==0.0.4
#virtualenv==1.8.4
#waitress==0.8.7
#wsgiref==0.1.2
#wstool==0.0.3


